requirements.txt
django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
psycopg2-binary==2.8.3
dj-database-url==0.5.0
django-filter=2.2.0